rename lib name to match class name
