+++ lib/kibana.rb
@@ -11,7 +11,7 @@ require 'yaml'
 require 'compat'
 require 'client_request'
 require 'id_request'
-require 'kransient_url'
+require 'ktransient_url'
 require 'query'
 require 'kelastic'
 require 'version'
