+++ lib/ktransient_url.rb
